VRT: fix integer overflow when dealing with buffer > 2GB (patch by patnichols, #5734)
git-svn-id: https://svn.osgeo.org/gdal/branches/1.11@27957 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/vrt/vrtsources.cpp
@@ -1673,7 +1673,7 @@ VRTComplexSource::RasterIO( int nXOff, int nYOff, int nXSize, int nYSize,
     return RasterIOInternal(nReqXOff, nReqYOff, nReqXSize, nReqYSize,
                        ((GByte *)pData)
                             + nPixelSpace * nOutXOff
-                            + nLineSpace * nOutYOff,
+                            + (size_t)nLineSpace * nOutYOff,
                        nOutXSize, nOutYSize,
                        eBufType,
                        nPixelSpace, nLineSpace );
